requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
banks==2.4.1
bcrypt==5.0.0
billiard==4.2.4
build==1.4.3
celery==5.4.0
certifi==2026.2.25
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.1
django-allauth==65.1.0
django-celery-beat==2.9.0
django-cors-headers==4.4.0
django-timezone-field==7.2.1
djangorestframework==3.15.2
durationpy==0.10
filelock==3.28.0
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
kubernetes==35.0.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.29
langchain-ollama==1.1.0
langchain-text-splitters==1.1.1
langsmith==0.7.31
llama-index-core==0.14.20
llama-index-embeddings-huggingface==0.7.0
llama-index-instrumentation==0.5.0
llama-index-llms-ollama==0.10.1
llama-index-workflows==2.19.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.4
openai==2.31.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
pybase64==1.4.3
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
Pygments==2.20.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
torch==2.11.0
opencv-python==4.10.0.84
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
psycopg[binary]==3.2.1
albumentations>=2.0.8
pycocotools>=2.0.11

